expenseApi/expense/requirements.txt
asgiref==3.5.2
asn1crypto==0.24.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
colorama==0.3.7
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distlib==0.3.5
distro-info===0.18ubuntu0.18.04.1
Django==4.1
djangorestframework==3.13.1
filelock==3.8.0
httplib2==0.9.2
idna==2.6
keyring==10.6.0
keyrings.alt==3.0
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lightdm-gtk-greeter-settings==1.2.2
netifaces==0.10.4
oauth==1.0.1
pexpect==4.2.1
Pillow==5.1.0
platformdirs==2.5.2
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
PyGObject==3.26.1
python-apt==1.6.5+ubuntu0.7
python-dateutil==2.6.1
python-debian==0.1.32
pytz==2022.2.1
pyxattr==0.6.0
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
sqlparse==0.4.2
ssh-import-id==5.7
ufw==0.36
urllib3==1.22
usb-creator==0.3.3
virtualenv==20.16.3
wadllib==1.3.2
xkit==0.0.0
youtube_dl==2018.3.14
zope.interface==4.3.2
